installer admin fix
ce igre ne zazenemo kot admin nas opozori in ne zazenemo igre

+++ igra.cpp
@@ -179,7 +179,7 @@ int GameManager::init() { // funkcija se klice le enkrat, ko se zazene SDL okno
 	}
 	*/
 
-	srand(time(NULL)); // poskrbi da generiramo nakljucna stevila
+	srand((unsigned int)time(NULL)); // poskrbi da generiramo nakljucna stevila
 	return EXIT_SUCCESS;
 }
